requirements.txt
plexapi==4.15.4
python-dotenv==1.0.0
colorama==0.4.6
tzdata == 2025.1